build-llvm.py: Add '--full-toolchain'
When doing upstream LLVM development or using the toolchain produced by
the script system-wide,it is advantageous to build as much of the code
as possible, rather than a subset like the script currently does.

Introduce '--full-toolchain', which stops the script from disabling
certain projects, targets, and configuration options. The '--projects',
'--targets', and '--defines' options are still respected so this
behavior can be more finely controlled.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>
Signed-off-by: KyuoFoxHuyu <88403766@qq.com>