build-llvm.py: Update GOOD_REVISION to b5983a38cbf4eb405fe9583ab89e15db1dcfa173
This has been qualified on aarch64 and x86_64 hosts using my
llvm-kernel-testing repository.

I had initially wanted to use [1] as the known good revision for this
upgrade (which is the parent of [2], which broke certain builds and is
still being dealt with on the Linux side) but that revision is
susceptible to warnings caused by [3], which was fixed by [4], which
occurs after [2]). There is even more breakage from a patch similar to
[2], [5], which is not yet resolved, so be conservative and stay further
back until it can be sorted out.

[1]: https://github.com/llvm/llvm-project/commit/99a1aeefb3d6be2018b591ed8c184c6f75fac386
[2]: https://github.com/llvm/llvm-project/commit/d81ce04587c006b6731198956c522c93d0df105
[3]: https://github.com/llvm/llvm-project/commit/0123deb3a6f0a83095287f51b07c77b7b43ab847
[4]: https://github.com/llvm/llvm-project/commit/844e9534c6d99ddb6bada740839760fa24d17cb6
[5]: https://github.com/llvm/llvm-project/commit/a79995ca6004082774a87f7a58ab6be5343364b7

Signed-off-by: Nathan Chancellor <nathan@kernel.org>