tc_build: llvm: Fix distribution_components with disabled projects
It is possible to use '--projects' to build certain combinations of
subprojects, such as just 'llvm' or 'clang' to build just LLVM or LLVM
and clang respectively. When either the 'clang' or 'lld' subproject is
not enabled, there is an error when running cmake due to their targets
being present in the distributions components list.

Only add them to distribution_components if the projects that contain
them are enabled.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>